feature: refactor query game by criteria with pipeline for more filter in the future
